requirements.txt
openai==1.84.0
dotenv== 0.9.9
redis==6.2.0
pypdf==5.6.0
numpy==2.0.2
black==25.1.0
openai-whisper==20240930
python-dotenv~=1.1.0
gradio~=4.44.1
pydantic==2.10.6